lab2 完成

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,7 +6,8 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        // int quotient = top / bottom;
+        float quotient = (float) top / bottom;
         int result = Math.round(quotient);
         return result;
     }
